% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_paraest_IME.R
\name{Cauchy_intervalparaest_IME_lambda}
\alias{Cauchy_intervalparaest_IME_lambda}
\title{Interval estimation of the positional parameter of a Cauchy distribution using R.}
\usage{
Cauchy_intervalparaest_IME_lambda(x, mu, C, alpha1, alpha2)
}
\arguments{
\item{x}{the i.i.d Cauchy variables (numeric)}

\item{mu}{the positional parameter of a Cauchy distribution (numeric)}

\item{C}{a suitable positive constant}

\item{alpha1}{alpha1+alpha2=alpha, in which alpha is the confidence coefficient of the confidence interval (numeric)}

\item{alpha2}{alpha1+alpha2=alpha, in which alpha is the confidence coefficient of the confidence interval (numeric)}
}
\value{
the interval estimation of the scale parameter \code{n}
}
\description{
Inverse moment interval estimation of the positional parameter of a Cauchy distribution if the scale parameter is known.
}
\examples{
\dontrun{
x<-rcauchy(50,0,1)
mu<-0
C<-1
alpha1<-0.025
alpha2<-0.025
Cauchy_intervalparaest_IME_lambda(x,mu,C,alpha1,alpha2)
}
}
