% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_paraest_IME.R
\name{Cauchy_intervalparaest_IME_mu}
\alias{Cauchy_intervalparaest_IME_mu}
\title{Interval estimation of the positional parameter of a Cauchy distribution using R.}
\usage{
Cauchy_intervalparaest_IME_mu(x, lambda, alpha)
}
\arguments{
\item{x}{the i.i.d Cauchy variables (numeric)}

\item{lambda}{the scale parameter of a Cauchy distribution (numeric)}

\item{alpha}{the confidence coefficient of the confidence interval (numeric)}
}
\value{
the interval estimation of the positional parameter \code{n}
}
\description{
Inverse moment interval estimation of the positional parameter of a Cauchy distribution if the scale parameter is known.
}
\examples{
\dontrun{
x<-rcauchy(50,0,1)
lambda<-1
alpha<-0.05
Cauchy_intervalparaest_IME_mu(x,lambda,alpha)
}
}
