% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_paraest_IME.R
\name{Cauchy_paraest_IME_lambda}
\alias{Cauchy_paraest_IME_lambda}
\title{Predict the scale parameter of a Cauchy distribution using R.}
\usage{
Cauchy_paraest_IME_lambda(x, mu, C)
}
\arguments{
\item{x}{the i.i.d Cauchy variables (numeric)}

\item{mu}{the positional parameter of a Cauchy distribution (numeric)}

\item{C}{a suitable positive constant}
}
\value{
the point estimation of the scale parameter \code{n}
}
\description{
Inverse moment point estimation of the scale parameter of a Cauchy distribution if the positional parameter is known.
}
\examples{
\dontrun{
x<-rcauchy(50,0,1)
mu<-0
C<-1
Cauchy_paraest_IME_lambda(x,mu,C)
}
}
