% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_paraest_IME.R
\name{Cauchy_paraest_IME_mu}
\alias{Cauchy_paraest_IME_mu}
\title{Predict the positional parameter of a Cauchy distribution using R.}
\usage{
Cauchy_paraest_IME_mu(x, lambda)
}
\arguments{
\item{x}{the i.i.d Cauchy variables (numeric)}

\item{lambda}{the scale parameter of a Cauchy distribution (numeric)}
}
\value{
the point estimation of the positional parameter \code{n}
}
\description{
Inverse moment point estimation of the positional parameter of a Cauchy distribution if the scale parameter is known.
}
\examples{
\dontrun{
x<-rcauchy(50,0,1)
lambda<-1
Cauchy_paraest_IME_mu(x,lambda)
}
}
