% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cauchy_paraest_mu}
\alias{Cauchy_paraest_mu}
\title{Predict the positional parameter of a Cauchy distribution using Rcpp.}
\usage{
Cauchy_paraest_mu(x, a)
}
\arguments{
\item{x}{the sorted i.i.d Cauchy variables (NumericVector)}

\item{a}{the method (1 or 2) (int)}
}
\value{
the point estimation of the positional parameter
}
\description{
Predict the positional parameter of a Cauchy distribution using Rcpp.
}
\examples{
\dontrun{
x<-rcauchy(50,0,1)
Cauchy_paraest_mu(x,1)
}
}
